Change: Move AnonymousCommandHandler to Game Server Script due to simplicity
