+++ GameServerScripts/GameServerScripts.csproj
@@ -102,6 +102,7 @@
       <SubType>Code</SubType>
     </Compile>
     <Compile Include="behaviour\examples\TestBehaviour.cs" />
+    <Compile Include="commands\player\anonymous.cs" />
     <Compile Include="commands\serverinfo.cs" />
     <Compile Include="commands\SwitchCommandHandler.cs" />
     <Compile Include="commands\version.cs" />
@@ -270,6 +271,7 @@
     <Compile Include="titles\TimeBasedTitle.cs" />
   </ItemGroup>
   <ItemGroup>
+    <Folder Include="commands\player" />
     <Folder Include="dbupdater" />
     <Folder Include="dbupdater\insert" />
     <Folder Include="dbupdater\replace" />
